PaimonsNotebook 1.5.0
### 新增

***

- **成就管理**

  - 成就管理遵循UIAF协议

- ##### 祈愿记录相关

  - 祈愿记录设置选项中增加了**关于UIGF**选项
  - 现在祈愿记录支持UIGF v3.0了
  - 现在可以在祈愿记录设置界面查看当前的UIGF版本了

- **设置相关**

  - 设置内新增了**快捷方式管理**的选项
    - 可将常用的功能配置在快捷方式中(Shortcuts)
    - 使用方法：**在桌面长按派蒙笔记本应用图标**唤起快捷方式列表
  - 设置内新增了**恢复默认设置**的选项
    - 点击后会将设置界面的所有选项重置为默认值
  - 设置内的检查更新可用了
    - 启动时会自动检查新版本(可关闭)

- ##### 文章详情

  - 现在可以通过点击文章详情内的**#话题名称#**或**底部的标签**跳转到对应的话题界面了
  - 话题文章列表中部分文章不会显示，这是有意的设置
    - 根据不同的话题，不会显示的内容约占总体的极小一部分，约1‰，或许比这个值更小
  - 话题文章列表总是显示20条数据，不会随着滑动列表获取更多数据，这是有意的设置

***

### 修改

- ##### 桌面组件

  - 调整了所有与每日任务相关的桌面组件的描述
    - 当已领取每日任务奖励显示**已完成**
    - 当未领取且完成的每日任务数量到达上限显示**待领取**
    - 当未达成每日任务数量显示**完成数/总数**(如2/4)

- ##### 祈愿记录

  - 调整了祈愿记录角色/武器界面的边距

- ##### 文章详情

  - 文章顶部的时间描述从具体的日期变更为模糊的描述
    - 如果文章的发布日期是最近的3天，会显示距离发布日期的模糊时间描述，如20分钟前、12小时前、昨天
    - 超过3天的显示具体的日期

***

### 修复

- ##### 祈愿记录

  - 修复输入祈愿记录URL完成按钮的逻辑错误

- ##### 元数据

  - 修复了部分情况下元数据更新失败的显示问题

  - 修复在缺少元数据的情况下进入角色/武器资料界面闪退的问题

- ##### 桌面组件配置

  - 修复桌面组件配置界面更改组件顶部关闭按钮部分被状态栏遮挡的问题
  - 修复在安卓13(API 33)使用桌面组件配置界面添加桌面组件异常的问题
  - 修复在安卓14(API 34)使用桌面组件配置界面添加桌面组件导致的闪退问题

- ##### 文章详情

  - 现在文章详情界面可以正确显示图片了
  - 修复了视频播放组件不能显示封面的问题
  - 修复文章详情界面中折叠内容中包含图片，展开折叠内容导致的崩溃
  - 修复了文章详情加载网页退出时，可能导致内存泄漏

- ##### 祈愿记录

  - 修复祈愿记录导入空数据导致的警告
  - 修复祈愿记录总览卡片显示的首次祈愿与最后一次祈愿日期错误的问题

+++ app/src/main/java/com/lianyi/paimonsnotebook/common/util/file/FileHelper.kt
@@ -1,7 +1,5 @@
 package com.lianyi.paimonsnotebook.common.util.file
 
-import android.Manifest
-import android.app.Activity
 import android.content.ContentResolver
 import android.media.MediaScannerConnection
 import android.net.Uri
@@ -11,7 +9,6 @@ import android.os.FileUtils
 import android.webkit.MimeTypeMap
 import com.lianyi.paimonsnotebook.common.application.PaimonsNotebookApplication
 import com.lianyi.paimonsnotebook.common.util.image.PaimonsNotebookImageLoader
-import com.lianyi.paimonsnotebook.common.util.system_service.SystemService
 import java.io.BufferedInputStream
 import java.io.File
 import java.io.FileOutputStream
@@ -25,39 +22,39 @@ object FileHelper {
         PaimonsNotebookApplication.context
     }
 
-    val privateStoragePath by lazy {
+    private val privateStoragePath by lazy {
         context.getExternalFilesDir(null)
     }
 
-    //如果有外部存储权限就存储到外部的downloads文件夹
-    private val rootPath by lazy {
-        Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS)
+    //如果有外部存储权限就存储到外部的documents文件夹
+    private val rootPath: File?
+        get() = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS)
             ?.resolve("PaimonsNotebook") ?: privateStoragePath
-    }
 
     //存储图片的路径
-    val saveImagePath by lazy {
-        rootPath?.resolve("image")!!
-    }
+    private val saveImagePath
+        get() =
+            rootPath?.resolve("image")!!
 
     //存储元数据的路径
-    val saveFileMetadataPath by lazy {
-        privateStoragePath?.resolve("metadata")!!
-    }
+    private val saveFileMetadataPath
+        get() =
+            privateStoragePath?.resolve("metadata")!!
 
-    //存储元数据图片的路径
-    val metadataImagesPath by lazy {
-        saveFileMetadataPath.resolve("images")
-    }
-
-    val saveFileUIGFPath by lazy {
-        privateStoragePath?.resolve("uigf")!!
-    }
+    //存储安装包的路径
+    val saveFilePackagePath
+        get() =
+            privateStoragePath?.resolve("package")!!
 
     //存储祈愿记录的路径
-    val saveFileGachaItemsPath by lazy {
-        rootPath?.resolve("gacha")!!
-    }
+    private val saveFileGachaItemsPath
+        get() =
+            rootPath?.resolve("gacha")!!
+
+    //存储成就记录的路径
+    private val saveFileAchievementsPath
+        get() =
+            rootPath?.resolve("achievements")!!
 
     /*
     * 从coil本地磁盘缓存中保存图片
@@ -151,22 +148,14 @@ object FileHelper {
         }
     }
 
-    fun getFile(rootPath: File, fileName: String): File? {
-        val file = File(rootPath, fileName)
-        return if (file.exists()) {
-            file
-        } else {
-            null
-        }
-    }
-
-    fun getSaveFile(rootPath: File, fileName: String): File {
+    fun getFile(rootPath: File, fileName: String): File {
         if (!rootPath.exists()) {
             rootPath.mkdirs()
         }
 
         return File(rootPath, fileName)
     }
+
 
     /*
     * 以流的方式保存文件
@@ -209,45 +198,45 @@ object FileHelper {
         }
     }
 
-
-    //获取元数据存储文件
-    fun getMetadataSaveFile(fileName: String): File {
-        if (!saveFileMetadataPath.exists()) {
-            saveFileMetadataPath.mkdirs()
+    //获取存储的文件
+    private fun getSaveFile(
+        rootPath: File,
+        fileName: String,
+        onExistsDelete: Boolean = true
+    ): File {
+        if (!rootPath.exists()) {
+            rootPath.mkdirs()
         }
 
-        return File(saveFileMetadataPath, "${fileName}.json")
-    }
+        val file = File(rootPath, fileName)
 
-
-    //获取uigf存储文件
-    fun getUIGFJsonSaveFile(name: String): File {
-        if (!saveFileGachaItemsPath.exists()) {
-            saveFileGachaItemsPath.mkdirs()
-        }
-
-        val file = File(saveFileGachaItemsPath, "${name}.json")
-
-        if (file.exists()) {
+        if (file.exists() && onExistsDelete) {
             file.delete()
         }
 
         return file
     }
 
-    //检查所需的权限
-    private fun checkPermission(): Boolean {
-        return SystemService.checkPermission(
-            Manifest.permission.READ_EXTERNAL_STORAGE,
-            Manifest.permission.WRITE_EXTERNAL_STORAGE
-        )
-    }
+    /*
+    * 获取元数据存储文件
+    * 验证元数据时也会用到该方法,onExistsDelete必须为false,否则会导致重复下载导致无法进入程序
+    * */
+    fun getMetadataSaveFile(fileName: String) =
+        getSaveFile(saveFileMetadataPath, "${fileName}.json", false)
 
-    private fun requestPermission(activity: Activity) = SystemService.requestPermission(
-        activity,
-        Manifest.permission.READ_EXTERNAL_STORAGE,
-        Manifest.permission.WRITE_EXTERNAL_STORAGE
-    )
+
+    //获取uigf存储文件
+    fun getUIGFJsonSaveFile(name: String) =
+        getSaveFile(saveFileGachaItemsPath, "${name}.json")
+
+    //获取uiaf存储文件
+    fun getUIAFJsonSaveFile(name: String) =
+        getSaveFile(saveFileAchievementsPath, "${name}.json")
+
+    //获取安装包存储文件
+    fun getPackageSaveFile(name: String) =
+        getSaveFile(saveFilePackagePath, "${name}.apk")
+
 
     private val df
         get() =
